requirements.txt
audio-recorder-streamlit==0.0.10
openai==1.55.3
google-api-core==1.34.1
google-auth==2.29.0
google-cloud-dialogflow==2.30.0
google-crc32c==1.6.0
googleapis-common-protos==1.63.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
